requirements.txt
langchain
openai
pypdf
faiss-cpu
sentence-transformers
streamlit
tiktoken
python-dotenv
transformers
torch
gpt4all
unstructured
pyttsx3